Improves player button layout algorithm
Refines the algorithm for determining the center point for player buttons on the tabletop.

The new algorithm considers tolerance for panel alignment and uses vertical and horizontal lines to identify potential center points.
It then selects the point closest to the center of the tabletop, ensuring better player button placement.

Also, this moves the background drawables and clip settings to the player button creation, and disables container clipping so they render correctly.